+++ app/components/Header.tsx
@@ -20,7 +20,7 @@ export default function Header() {
       <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
         <div className="flex justify-between items-center h-16">
           <Link href="/" className="tap-target inline-flex items-center text-2xl font-bold text-primary-token hover:text-secondary-token transition-colors">
-            工事中
+            福大ピアプロ
           </Link>
 
           <nav className="hidden md:flex space-x-8">
