+++ components/ActivityDetailModal.tsx
@@ -86,14 +86,14 @@ export default function ActivityDetailModal({ activity, onClose }: ActivityDetai
         <button
           type="button"
           onClick={onClose}
-          className="absolute -right-1 -top-1 z-[110] flex h-10 w-10 items-center justify-center rounded-full bg-white text-gray-700 shadow-lg ring-1 ring-gray-200 transition hover:bg-gray-50 md:right-0 md:top-0"
+          className="tap-target absolute -right-1 -top-1 z-[110] flex h-11 w-11 items-center justify-center rounded-full bg-[var(--surface-raised)] text-primary-token shadow-lg ring-1 ring-[var(--border-subtle)] transition hover:bg-[var(--surface-muted)] md:right-0 md:top-0"
           aria-label="閉じる"
         >
           <svg className="h-5 w-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
             <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
           </svg>
         </button>
-        <div className="max-h-[min(92dvh,920px)] overflow-y-auto overflow-x-hidden overscroll-y-contain rounded-3xl bg-white shadow-sm ring-1 ring-gray-100">
+        <div className="max-h-[min(92dvh,920px)] overflow-y-auto overflow-x-hidden overscroll-y-contain rounded-3xl bg-[var(--surface-raised)] shadow-sm ring-1 ring-[var(--border-subtle)]">
           <ActivityDetailView activity={activity} />
         </div>
       </div>
